+++ app/models/WorkDiary.java
@@ -174,7 +174,7 @@ public class WorkDiary extends Model {
     /**
      * 列数
      */
-    public int retusu;
+    public double retusu;
     /**
      * 使用苗枚数
      */
